+++ test/snippet/alphabet/views/translate_usage.cpp
@@ -13,26 +13,26 @@ int main()
     // == [[T,Y,V,R],[R,T,Y,V],[V,R,T],[Y,V,R,T],[T,Y,V,R],[R,T,Y]]
 
     // single frame translation
-    auto v2 = vec | seqan3::views::translate(seqan3::translation_frames::FWD_FRAME_0);
+    auto v2 = vec | seqan3::views::translate(seqan3::translation_frames::forward_frame0);
     // == [[T,Y,V,R]]
 
     // reverse translation
-    auto v3 = vec | seqan3::views::translate(seqan3::translation_frames::FWD_REV_0);
+    auto v3 = vec | seqan3::views::translate(seqan3::translation_frames::forward_reverse0);
     // == [[T,Y,V,R],[Y,V,R,T]]
 
     // forward frames translation
-    auto v4 = vec | seqan3::views::translate(seqan3::translation_frames::FWD);
+    auto v4 = vec | seqan3::views::translate(seqan3::translation_frames::forward_frames);
     // == [[T,Y,V,R],[R,T,Y,V],[V,R,T]]
 
     // six frame translation
-    auto v5 = vec | seqan3::views::translate(seqan3::translation_frames::SIX_FRAME);
+    auto v5 = vec | seqan3::views::translate(seqan3::translation_frames::six_frames);
     // == [[T,Y,V,R],[R,T,Y,V],[V,R,T],[Y,V,R,T],[T,Y,V,R],[R,T,Y]]
 
     // function syntax
-    auto v6 = seqan3::views::translate(vec, seqan3::translation_frames::FWD_REV_0);
+    auto v6 = seqan3::views::translate(vec, seqan3::translation_frames::forward_reverse0);
     // == [[T,Y,V,R],[Y,V,R,T]]
 
     // combinability
-    auto v7 = vec | seqan3::views::complement | seqan3::views::translate(seqan3::translation_frames::FWD_REV_0);
+    auto v7 = vec | seqan3::views::complement | seqan3::views::translate(seqan3::translation_frames::forward_reverse0);
     // == [[C,M,H,A],[M,H,A,C]]
 }